Fix appdb circular imports.
